+++ doc/changelog.rst
@@ -19,6 +19,7 @@ What's New
 
   * Fixed ``ConnectionPool`` ``max_host_count`` argument not used.
   * Moved document scraping concerns from ``WebProcessorSession`` to ``ProcessingRule``.
+  * Renamed ``SSLVerficationError`` to ``SSLVerificationError``.
 
 
 0.1001.2 (2014-10-25)